Reorganize project repo strucutre
