+++ seeds/books_favorites.js
@@ -1,9 +1,9 @@
 /**
  * @param { import("knex").Knex } knex
- * @returns { Promise<void> } 
+ * @returns { Promise<void> }
  */
-const booksData = require('../../data/filtered-books-data');
-const favorites = require("../../data/favorites-data");
+const booksData = require('../data/filtered-books-data');
+const favorites = require("../data/favorites-data");
 
 exports.seed = function(knex) {
   // Deletes ALL existing entries
