+++ server.js
@@ -3,7 +3,7 @@ const express = require('express');
 const favorites = require('./data/favorites-data');
 const filteredBooks = require('./data/filtered-books-data');
 const app = express();
-const queries = require('./db/queries');
+const queries = require('./queries');
 app.use(express.json());
 
 
@@ -27,7 +27,7 @@ app.get('/api/v1/books/:id', (request, response) => {
         response.status(404).json({
           error: `Could not find book with id ${request.params.id}`
         });
-      } 
+      }
     })
     .catch(error => response.status(500).json({ error }));
 });
